% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{png.qqman}
\alias{png.qqman}
\alias{png.qqman-package}
\title{Create Q-Q and manhattan plots for GWAS data.}
\description{
A package for creating Q-Q and manhattan plots for GWAS data. See the package
vignette for details: \cr\cr
\code{vignette("png.qqman")}
}
\author{
Stephen Turner <\url{http://stephenturner.us}>
}
